Update 01_intro.ipynb
Minor modifications to Section 1.7 to clean up the recipe for installing the necessary Python environment for working through the exercises included in the BookNLP course.

To provide visual clarity between the commands to be entered on the Terminal prompt and the recipe text, the Markdown backslash character is used to format these commands. If the publishing server that renders these pages does not support this Markdown tagging, these backslashes will need to be removed.